Post detail Use Cases

+++ lib/feature/post_by_id/data/data_sources/comment_api_services.dart
@@ -10,5 +10,5 @@ sealed class CommentApiServices {
 
   @GET('/comments')
   Future<HttpResponse<List<CommentModel>>> getComments(
-      {@Query("postId") required double postId});
+      {@Query("postId") required int postId});
 }
